changed int to long

+++ ipc_shared_memory/include/vectorMatch.hpp
@@ -28,7 +28,7 @@ public:
 
     using shmKeyPair = struct{
         float dist;
-        int lineNum;
+        long lineNum;
     };
